# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReliAPI",
    "description": "Reliability layer for API calls: retries, caching, dedup, circuit breakers, LLM proxy with cost control.",
    "version": "1.1.0",
    "contact": {
      "name": "KikuAI-Lab",
      "url": "https://github.com/kiku-jw/reliapi",
      "email": "dev@kikuai.dev"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "paths": {
...
    "/paddle/checkout": {
      "post": {
        "tags": [
          "paddle"
        ],
        "summary": "Create Checkout",
        "description": "Create a Paddle checkout session for subscription.\n\nThis endpoint creates a checkout session through Paddle API.\nReturns checkout URL that user can redirect to.",
        "operationId": "create_checkout_paddle_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Checkout Paddle Checkout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paddle/subscription/{subscription_id}": {
      "get": {
        "tags": [
          "paddle"
        ],
        "summary": "Get Subscription",
        "description": "Get subscription status from Paddle.\n\nFetches subscription details from Paddle API.",
        "operationId": "get_subscription_paddle_subscription__subscription_id__get",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paddle/webhook": {
      "post": {
        "tags": [
          "paddle"
        ],
        "summary": "Paddle Webhook",
        "description": "Handle Paddle webhook events.\n\nProcesses webhook events from Paddle:\n- subscription.created\n- subscription.updated\n- subscription.cancelled\n- transaction.completed\n- transaction.payment_failed\n\nAll events are processed automatically without manual intervention.",
        "operationId": "paddle_webhook_paddle_webhook_post",
        "parameters": [
          {
            "name": "Paddle-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Paddle-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Paddle Webhook Paddle Webhook Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCheckoutRequest": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "customer_email": {
            "type": "string",
            "title": "Customer Email"
          },
          "success_url": {
            "type": "string",
            "title": "Success Url"
          },
          "cancel_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "customer_email",
          "success_url"
        ],
        "title": "CreateCheckoutRequest",
        "description": "Request to create a Paddle checkout."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionStatus": {
        "properties": {
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "cancel_at_period_end": {
            "type": "boolean",
            "title": "Cancel At Period End",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "subscription_id",
          "status",
          "plan_id"
        ],
        "title": "SubscriptionStatus",
        "description": "Subscription status response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}